requirements.txt
pydantic
langchain-core
langchain
langchain-community
langchain-openai
python-dotenv
requests
beautifulsoup4
faiss-cpu
simple_salesforce